Add button in espace responsable + view to remove OTP app from aidant
