+++ aidants_connect_web/urls.py
@@ -199,6 +199,11 @@ urlpatterns = [
         name="espace_responsable_aidant_add_app_otp",
     ),
     path(
+        "espace-responsable/aidant/<int:aidant_id>/supprimer-otp-app/",
+        espace_responsable.RemoveAppOTPFromAidant.as_view(),
+        name="espace_responsable_aidant_remove_app_otp",
+    ),
+    path(
         "espace-responsable/aidant/<int:aidant_id>/changer-organisations/",
         espace_responsable.change_aidant_organisations,
         name="espace_responsable_aidant_change_organisations",
